bridge/requirements.txt
python-qpid-proton>=0.27.0
